requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
openai
python-dotenv
jinja2
pydantic